d98bd0d4-7247-407a-b009-4cd0525a224b/ai-voice-detection/requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.24.0
